Enable initial read/write to registers
For writing: the address, value, length and 0x1f are written to a DMA segment.
The physical address of this DMA segment is then written to a specific SUME
register. For reading, you first 'write' using 0x00 and the address of the
register to read from and wait for the SUME to write the response to a DMA
segment for receiving.
There is also a rtag which can be used to check if we caught the right
transaction.

This seems simple enough to do without interrupt handlers or locking (without
any other transactions) and it works for now but the next step should be to
work on implementing those mechanisms and prepare for packet RX/TX.