Add basic packet RX/TX
Packets coming from SUME are forwarded to the OS and vice-versa. The
DMA transactions are based on the Linux driver: using only a single
"bouncebuffer" without any descriptors, then copying data to/from the
mbuf.

I'm not sure whether locking is implemented correctly but this mostly
works ok (for smaller packets at least).