Fix module register RX
Usually, an RX interrupt sends us vectors for RXBUF and RXDONE at the same time,
but sometimes, when the device is busy with multiple RX transactions (both
packet and modreg ones), the RX vectors can arrive separately. This could mean
that our modreg function can get timeout waiting for the READ state, and get it
later: after resetting the FSM to IDLE state.
So when we get the RXDONE vector in the IDLE state, we must read the value in
order for the HW to proceed to the next state and don't get stuck. We do this in
recovery.