Resolve an RX bug
When we get overflown with RX packets and reset the HW at the same time,
we can get stuck in the wrong RX state. This patch resets the RX state.
Add an additional check for port number so we don't crash if we get a
bogus packet in that case.